// suivi_de_stage_backend/src/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  STAGE_AFFECTE = 'STAGE_AFFECTE',
  STAGE_MODIFIE = 'STAGE_MODIFIE',
  STAGE_TERMINE = 'STAGE_TERMINE',
  FIN_STAGE_PROCHE = 'FIN_STAGE_PROCHE',
  EVALUATION = 'EVALUATION',
  INFORMATION = 'INFORMATION',
  RAPPORT_DEPOSE = 'RAPPORT_DEPOSE',
  RAPPORT_REVU = 'RAPPORT_REVU',
}